Move stuff common to vertex/fragment_program into the base class, including: Instructions, InputsRead, OutputsWritten, and Parameters. Also, added debug functions: _mesa_print_instruction(), _mesa_print_program_parameters() and revamp _mesa_print_program().

+++ src/mesa/drivers/dri/r300/r300_context.h
@@ -598,7 +598,7 @@ struct r300_vertex_program {
 	struct r300_vertex_shader_fragment params;
 	
 	int pos_end;
-	unsigned long num_temporaries; /* Number of temp vars used by program */
+	int num_temporaries; /* Number of temp vars used by program */
 	int inputs[VERT_ATTRIB_MAX];
 	int outputs[VERT_RESULT_MAX];
 };
